datasource/cx-analytics/requirements.txt
psycopg2-binary>=2.9.0
faker>=20.0.0
boto3>=1.28.0
pandas>=2.0.0
pyarrow>=14.0.0
